[ET-VK] 3/n Split dispatches between multiple command buffers. Track previous semaphore in context.
Pull Request resolved: https://github.com/pytorch/executorch/pull/12523

This diff is the third part of a series of diffs aiming to split dispatches between multiple command buffers. In this diff, we are tracking the previous semaphore in the context.

A new member variable `prev_semaphore_` was added to the `Context` class. This variable is used to store the semaphore of the previously submitted command buffer.
ghstack-source-id: 296448617

Differential Revision: [D78360037](https://our.internmc.facebook.com/intern/diff/D78360037/)